# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/jwt/reservation/create": {
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Crear una nueva reservación",
                "description": "Crea una nueva reservación para el vehículo especificado y opcionalmente configura un recordatorio.",
                "operationId": "97512533f516a348b7e9186b41bcf4a6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reservationDate",
                                    "carId"
                                ],
                                "properties": {
                                    "reservationDate": {
                                        "description": "Fecha y hora de la reservación en formato d-m-Y H:i",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-10-15 14:30"
                                    },
                                    "carId": {
                                        "description": "ID del vehículo para la reservación",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "wantReminder": {
                                        "description": "Si el usuario desea un recordatorio",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "typeReminder": {
                                        "description": "Tipo de recordatorio si se solicita",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "mechanicId": {
                                        "description": "ID del mecánico a notificar",
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reservación creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "reservation": {
                                                    "type": "object"
                                                },
                                                "reminder": {
                                                    "type": "string",
                                                    "example": "Sera enviado a traves de correo email (Por defecto)"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "reservation created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "description": "Detalles de los errores de validación",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}